Removing old AudioPlayer, replaced with AudioPlayer2
