+++ Tests/AudioKitTests/Node Tests/Generator Tests/AudioPlayerTests.swift
@@ -4,29 +4,36 @@ import XCTest
 
 class AudioPlayerTests: XCTestCase {
     // Because SPM doesn't support resources yet, render out a test file.
-    func generateTestFile() -> URL {
+    func generateTestFile() -> URL? {
         let osc = Oscillator()
         let engine = AudioEngine()
         engine.output = osc
         osc.start()
 
-        let mgr = FileManager.default
-        let url = mgr.temporaryDirectory.appendingPathComponent("test.aiff")
-        try? mgr.removeItem(at: url)
+        let url = FileManager.default.temporaryDirectory.appendingPathComponent("test.aiff")
+        try? FileManager.default.removeItem(at: url)
+
         var audioFormatSettings = Settings.audioFormat.settings
         audioFormatSettings["AVLinearPCMIsNonInterleaved"] = false
-        let file = try! AVAudioFile(forWriting: url, settings: audioFormatSettings)
 
-        try! engine.renderToFile(file, duration: 1)
-        print("rendered test file to \(url)")
+        do {
+            let file = try AVAudioFile(forWriting: url, settings: audioFormatSettings)
+            try engine.renderToFile(file, duration: 1)
+            Log("rendered test file to", url)
+        } catch let error as NSError {
+            Log(error, type: .error)
+            return nil
+        }
 
         return url
     }
 
     func testBasic() {
-        let url = generateTestFile()
-
-        let file = try! AVAudioFile(forReading: url)
+        guard let url = generateTestFile(),
+            let file = try? AVAudioFile(forReading: url) else {
+            XCTFail("Didn't generate test file")
+            return
+        }
 
         let engine = AudioEngine()
         let player = AudioPlayer()
@@ -43,9 +50,11 @@ class AudioPlayerTests: XCTestCase {
     }
 
     func testLoop() {
-        let url = generateTestFile()
-        let file = try! AVAudioFile(forReading: url)
-        let buffer = try! AVAudioPCMBuffer(file: file)!
+        guard let url = generateTestFile(),
+            let buffer = try? AVAudioPCMBuffer(url: url) else {
+            XCTFail("Couldn't create buffer")
+            return
+        }
 
         let engine = AudioEngine()
         let player = AudioPlayer()
@@ -63,17 +72,11 @@ class AudioPlayerTests: XCTestCase {
         // audition(audio)
     }
 
-    func testScheduleEarly() {
-        let url = generateTestFile()
-
-        let file = try! AVAudioFile(forReading: url)
-
-        let player = AudioPlayer()
-        player.scheduleFile(file, at: nil)
-    }
-
     func testScheduleFile() {
-        let url = generateTestFile()
+        guard let url = generateTestFile() else {
+            XCTFail("Didn't generate test file")
+            return
+        }
 
         let engine = AudioEngine()
         let player = AudioPlayer()
@@ -81,22 +84,20 @@ class AudioPlayerTests: XCTestCase {
         engine.output = player
 
         let audio = engine.startTest(totalDuration: 2.0)
-        var output = ""
-        player.scheduleFile(url, at: nil, options: .loops) {
-            output = "success"
-        }
+        player.scheduleBuffer(url: url, at: nil, options: .loops)
         player.play()
         audio.append(engine.render(duration: 2.0))
         engine.stop()
 
         testMD5(audio)
-        XCTAssertEqual(output, "success")
     }
 
     func testVolume() {
-        let url = generateTestFile()
-
-        let file = try! AVAudioFile(forReading: url)
+        guard let url = generateTestFile(),
+            let file = try? AVAudioFile(forReading: url) else {
+            XCTFail("Couldn't create file")
+            return
+        }
 
         let engine = AudioEngine()
         let player = AudioPlayer()
